+++ src/main/groovy/com/moowork/gradle/grunt/GruntTask.groovy
@@ -23,33 +23,46 @@ class GruntTask
                 "Grunt-CLI not installed in node_modules, please first run 'gradle ${GruntPlugin.GRUNT_INSTALL_NAME}'" )
         }
 
-        // If colors are disabled, add --no-color to args
-        if ( !this.project.grunt.colors ) {
+        def addArg = { List<?> original, def newArg ->
             def tempArgs = []
-            tempArgs.addAll( args as List )
-            tempArgs.add( '--no-color' )
-            setArgs( tempArgs )
+            tempArgs.addAll( original )
+            tempArgs.add( newArg )
+            return tempArgs
+        }
+
+        // On multi project scenario, need to specify Gruntfile.js location
+        setArgs( addArg( args as List, "--gruntfile ${this.project.grunt.workDir}/Gruntfile.js" ) )
+
+        // If colors are disabled, add --no-color to args
+        if ( !this.project.grunt.colors )
+        {
+            setArgs( addArg( args as List, '--no-color' ) )
         }
 
         // If output should be buffered (useful when running in parallel)
         // set standardOutput of ExecRunner to a buffer
         def bufferedOutput
-        if ( this.project.grunt.bufferOutput ) {
+        if ( this.project.grunt.bufferOutput )
+        {
             bufferedOutput = new ByteArrayOutputStream()
-            setExecOverrides({
+            setExecOverrides( {
                 it.standardOutput = bufferedOutput
-            })
+            } )
         }
 
         setWorkingDir( this.project.grunt.workDir )
         setScript( localGrunt )
 
         // If the exec fails, make sure to still print output
-        try {
+        try
+        {
             super.exec()
-        } finally {
+        }
+        finally
+        {
             // If we were buffering output, print it
-            if ( this.project.grunt.bufferOutput ) {
+            if ( this.project.grunt.bufferOutput )
+            {
                 println "Output from ${this.project.grunt.workDir}/Gruntfile.js"
                 println bufferedOutput.toString()
             }