+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.13 *(not released)*
 * Bumped gradle wrapper version to 2.14
 * Updated plugin versions and test versions
 * Using gradle-node-plugin version 0.13
+* Fix multi gradle project Gruntfile.js location _(mhmxs)_
 * ...
 
 Version 0.12 *(2016-03-10)*